services/api/requirements.txt
cryptography
db==0.1.1
Flask==2.0.2
flask_cors
flask-marshmallow==0.14.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
marshmallow==3.14.1
marshmallow-sqlalchemy==0.27.0
python-dotenv==0.19.2
PyMySQL==0.10.0
Flask-MySQL==1.5.1
Flask-RESTify==0.1.3
Flask-SSLify==0.1.5
psycopg2==2.8.6
